Changed Telegram Bot API version to 5.7.1

+++ src/main/java/datasourse/repositories/BotChatRepository.java
@@ -14,7 +14,7 @@ public interface BotChatRepository extends JpaRepository<BotChat, Integer> {
     @Modifying
     @Transactional
     @Query("UPDATE BotChat chat SET chat.users = ?2 WHERE chat.id = ?1")
-    void updateUserName(int id, Map<Integer, ChatUser> users);
+    void updateUserName(int id, Map<Long, ChatUser> users);
 
     @Query("SELECT SUM(chat.callCounter) FROM BotChat chat")
     int getSumOfCallCounters();
